fix: qwen35 tool parsing and response fix
