+++ crates/core/src/provider/gguf/mod.rs
@@ -5,5 +5,7 @@ pub use model::GgufBaseModel;
 mod checkpoint;
 mod model;
 mod template;
+mod tool;
 
-pub use template::{ToolCallParser, apply_chat_template};
+pub use template::apply_chat_template;
+pub use tool::{TemplatePatterns, ToolCallParser, get_tool_call_regexes};
